Moved DxilEraseDeadRegions to after DxilRemoveDeadBlocks
